+++ src/Modbus/Buffer.cpp
@@ -1,5 +1,5 @@
 /*
- * MPU simulator.
+ * Implements generic Modbus Buffer functions.
  *
  * Developed for the Vera C. Rubin Observatory Telescope & Site Software Systems.
  * This product includes software developed by the Vera C.Rubin Observatory Project
@@ -20,24 +20,41 @@
  * this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-#ifndef CRIO_SIMULATEDMPU_H_
-#define CRIO_SIMULATEDMPU_H_
+#include <spdlog/spdlog.h>
 
-#include <cRIO/MPU.h>
+#include <Modbus/Buffer.h>
+#include <Modbus/CRC.h>
 
-namespace LSST {
-namespace cRIO {
+using namespace Modbus;
 
-class SimulatedMPU : public MPU {
-public:
-    SimulatedMPU() : MPU(0, 0) {}
-    SimulatedMPU(uint16_t* buffer, size_t length) : MPU(0, 0) { setBuffer(buffer, length); }
+Exception::Exception(uint8_t address, uint8_t func, uint8_t exception)
+        : std::runtime_error(fmt::format(
+                  "ModBus Exception {2} (ModBus address {0}, ModBus response function {1} (0x{1:02x})).",
+                  address, func, exception)) {}
 
-    void loopWrite() override {}
-    void loopRead(bool timedout) override {}
-};
+Buffer::Buffer() : std::vector<uint8_t>() {}
 
-}  // namespace cRIO
-}  // namespace LSST
+Buffer::~Buffer() {}
 
-#endif  // CRIO_SIMULATEDMPU_H_
+uint16_t Buffer::getCalcCrc() {
+    CRC crc(*this);
+    return crc.get();
+}
+
+void Buffer::callFunction(uint8_t address, uint8_t func) {
+    write(address);
+    write(func);
+    writeCRC();
+}
+
+void Buffer::writeI24(int32_t data) {
+    pushBuffer(data >> 16);
+    pushBuffer(data >> 8);
+    pushBuffer(data);
+}
+
+void Buffer::writeCRC() {
+    uint16_t crc = getCalcCrc();
+    pushBuffer(crc & 0xFF);
+    pushBuffer((crc >> 8) & 0xFF);
+}