+++ tests/TestFPGA.h
@@ -30,12 +30,13 @@
 #include <cRIO/FPGA.h>
 #include <cRIO/PrintILC.h>
 #include <cRIO/SimulatedILC.h>
+#include <ILC/ILCBusList.h>
 
 enum FPGAAddress { MODBUS_A_RX = 21, MODBUS_A_TX = 25, HEARTBEAT = 62 };  // namespace FPGAAddress
 
 class TestILC : public LSST::cRIO::PrintILC {
 public:
-    TestILC(uint8_t bus) : LSST::cRIO::PrintILC(bus) {}
+    TestILC(uint8_t bus) : ILC::ILCBusList(bus), LSST::cRIO::PrintILC(bus) {}
 
 protected:
     void processChangeILCMode(uint8_t address, uint16_t mode) override;
@@ -52,7 +53,7 @@ public:
     uint16_t getTxCommand(uint8_t bus) override { return FPGAAddress::MODBUS_A_TX; }
     uint16_t getRxCommand(uint8_t bus) override { return FPGAAddress::MODBUS_A_RX; }
     uint32_t getIrq(uint8_t bus) override { return 1; }
-    void writeMPUFIFO(LSST::cRIO::MPU&) override {}
+    void writeMPUFIFO(uint8_t bus, std::vector<uint8_t> mpu_data) override { last_mpu = mpu_data; }
     std::vector<uint8_t> readMPUFIFO(LSST::cRIO::MPU&) override {
         return std::vector<uint8_t>({0xff, 0x0fe});
     }
@@ -64,6 +65,8 @@ public:
 
     void setPages(uint8_t* pages) { _pages = pages; }
     void setSimulatedIRQs(uint32_t irqs) { _simulatedIRQs = irqs; }
+
+    std::vector<uint8_t> last_mpu;
 
 protected:
     void processServerStatus(uint8_t address, uint8_t mode, uint16_t status, uint16_t faults) override;